+++ src/app/component/todo-list/todo-list.component.ts
@@ -1,9 +1,11 @@
 import {
   Component,
   OnInit
-}                    from '@angular/core';
-import {TodoService} from "../../service/todo.service";
-import {Todo}        from "../../model/Todo";
+}                               from '@angular/core';
+import {TodoService}            from "../../service/todo.service";
+import {Todo}                   from "../../model/Todo";
+import {MatDialog}              from "@angular/material/dialog";
+import {DefaultDialogComponent} from "../default-dialog/default-dialog.component";
 
 @Component({
   selector:    'app-todo-list',
@@ -14,7 +16,7 @@ export class TodoListComponent implements OnInit {
 
   todoList: Todo[] = [];
 
-  constructor(private todoService: TodoService) {
+  constructor(private todoService: TodoService, private dialog: MatDialog) {
   }
 
   ngOnInit() {
@@ -28,6 +30,12 @@ export class TodoListComponent implements OnInit {
   }
 
   delete(todo: Todo) {
-    this.todoService.deleteTodo(todo.id).subscribe(() => this.getTodoList());
+    let dialog = this.dialog.open(DefaultDialogComponent, {
+      width: '250px',
+      data:  {title: "Todoを削除しますか？"}
+    });
+    dialog.afterClosed().subscribe((result) => {
+      if (result) this.todoService.deleteTodo(todo.id).subscribe(() => this.getTodoList());
+    })
   }
 }
